assb/requirements.txt
numpy>=1.21.0,<2.0
pandas>=1.3.0,<2.0
scikit-learn>=0.24.0,<2.0
matplotlib>=3.4.0,<4.0
requests>=2.25.0,<3.0
# Optional dependencies
vaderSentiment>=3.3.2  # For sentiment analysis
ccxt>=2.2.2 # exchange binding
python-dotenv
tweepy
transformers



